[modules/ipmb] move board specific code to port/board/afc-common
