+++ port/board/afc-common/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(BOARD_PATH ${CMAKE_CURRENT_SOURCE_DIR})
 
 set(PROJ_SRCS ${PROJ_SRCS} ${BOARD_PATH}/board_led.c)
+set(PROJ_SRCS ${PROJ_SRCS} ${BOARD_PATH}/board_ipmb.c)
 
 set(PROJ_HDRS ${PROJ_HDRS} ${BOARD_PATH})
 set(PROJ_SRCS ${PROJ_SRCS} PARENT_SCOPE)
